Get it running again; better search for curses library
Uses the LibraryCheck module to check for the best curses library
available. This should work on Linux, BSD, or OS X.

Specify the library version; this is required on Redhat and others,
because 'libfoo.so' is a GNU ld script (like "INPUT(-lfoo)") that
dlopen doesn't handle. This is good for portability and to avoid a
"DLL Hell" of underspecified dependencies, but may stretch the
simplistic search implemented here. How portable this is remains to be
seen, but it can't be worse than it was before.

Since DESTROY is only called during garbage collection, it's not really
appropriate for calling endwin(). So just document the user calling
endwin() directly, from a phaser.

When loading a symbol from the program's name space (libc, etc.), just
use `is native` with no arguments.

The Proxy usage changed (needs return-rw instead of return; or, even
better, just sink it out of the method.